Code refactor
Updating files after renaming models. Minor updates to email templates.
