+++ src/services/auth.service.js
@@ -4,7 +4,7 @@ const Promise = require('bluebird')
 const mongoose = require('mongoose')
 mongoose.Promise = Promise
 
-const User = require('../models/user.server.model')
+const User = require('../models/user.model')
 const tokenLife = config.TOKEN_LIFE
 const _ = require('lodash')
